Protocol cleanups
Remove TOCLIENT_ACCESS_DENIED_LEGACY and DenyAccess_Legacy
Remove deprecated GENERIC_CMD_UPDATE_NAMETAG_ATTRIBUTES
 -> Raise client initialization data version
Remove health and breath HUD compatibility code

u16 client damage, u16 hp, s32 server side hp change

+++ src/network/serverpackethandler.cpp
@@ -776,7 +776,7 @@ void Server::handleCommand_ChatMessage(NetworkPacket* pkt)
 
 void Server::handleCommand_Damage(NetworkPacket* pkt)
 {
-	u8 damage;
+	u16 damage;
 
 	*pkt >> damage;
 
@@ -812,7 +812,7 @@ void Server::handleCommand_Damage(NetworkPacket* pkt)
 				<< std::endl;
 
 		PlayerHPChangeReason reason(PlayerHPChangeReason::FALL);
-		playersao->setHP(playersao->getHP() - damage, reason);
+		playersao->setHP((s32)playersao->getHP() - damage, reason);
 		SendPlayerHPOrDie(playersao, reason);
 	}
 }
